move AzureExceptionHandler and RestExceptionHandlerInterceptor to package `common.exception`
